move io to core
