+++ backend/management/load_data.py
@@ -9,7 +9,7 @@ from app.utils import (
     add_extent_to_districtrmap,
     create_spatial_index,
 )
-from app.io import get_local_or_s3_path
+from app.core.io import get_local_or_s3_path
 from app.main import get_session
 from app.core.config import settings
 from functools import wraps
